Removed a lot of compiler warnings under Windows.
Some unit tests still fail. I'm not sure why.
Also, the visual studio debugger warns of an std::bad_weak_ptr
exception being thrown. I'm not sure how that can happen. It
probably has something to do with the Entity class.

+++ Examples/DirectionalLight.cpp
@@ -60,8 +60,8 @@ private:
 	virtual void onRenderUpdate() final
 	{
 		using namespace gintonic;
-		mCubeEntity->postMultiplyRotation(quatf::axis_angle(vec3f(1.0f, 0.0f, 0.0f), mDeltaTime / 10.0f));
-		mCubeEntity->postMultiplyRotation(quatf::axis_angle(vec3f(0.0f, 1.0f, 0.0f), mDeltaTime / 5.0f));
+		mCubeEntity->postMultiplyRotation(quatf::axis_angle(vec3f(1.0f, 0.0f, 0.0f), static_cast<float>(mDeltaTime) / 10.0f));
+		mCubeEntity->postMultiplyRotation(quatf::axis_angle(vec3f(0.0f, 1.0f, 0.0f), static_cast<float>(mDeltaTime) / 5.0f));
 	}
 
 };
